Remove kyma upgrade capability
