+++ internal/orchestration/handlers/orchestration_handler_test.go
@@ -37,7 +37,7 @@ func TestStatusHandler_AttachRoutes(t *testing.T) {
 		require.NoError(t, err)
 
 		logs := logrus.New()
-		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, nil, 100, logs)
+		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, 100, logs)
 
 		req, err := http.NewRequest("GET", "/orchestrations?page_size=1", nil)
 		require.NoError(t, err)
@@ -154,7 +154,7 @@ func TestStatusHandler_AttachRoutes(t *testing.T) {
 		require.NoError(t, err)
 
 		logs := logrus.New()
-		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, nil, 100, logs)
+		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, 100, logs)
 
 		urlPath := fmt.Sprintf("/orchestrations/%s/operations", fixID)
 		req, err := http.NewRequest("GET", urlPath, nil)
@@ -223,7 +223,7 @@ func TestStatusHandler_AttachRoutes(t *testing.T) {
 		require.NoError(t, err)
 
 		logs := logrus.New()
-		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, nil, 100, logs)
+		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, 100, logs)
 
 		urlPath := fmt.Sprintf("/orchestrations/%s/operations", fixID)
 		req, err := http.NewRequest("GET", urlPath, nil)
@@ -275,7 +275,7 @@ func TestStatusHandler_AttachRoutes(t *testing.T) {
 		require.NoError(t, err)
 
 		logs := logrus.New()
-		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, nil, 100, logs)
+		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, 100, logs)
 
 		req, err := http.NewRequest("PUT", fmt.Sprintf("/orchestrations/%s/cancel", fixID), nil)
 		require.NoError(t, err)
@@ -379,7 +379,7 @@ func TestStatusHandler_AttachRoutes(t *testing.T) {
 		require.NoError(t, err)
 
 		logs := logrus.New()
-		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, nil, 100, logs)
+		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, 100, logs)
 
 		urlPath := fmt.Sprintf("/orchestrations/%s/operations", orchestration1ID)
 		req, err := http.NewRequest("GET", urlPath, nil)
@@ -479,7 +479,7 @@ func TestStatusRetryHandler_AttachRoutes(t *testing.T) {
 
 		logs := logrus.New()
 		clusterQueue := process.NewQueue(&testExecutor{}, logs)
-		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, clusterQueue, 100, logs)
+		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), clusterQueue, 100, logs)
 
 		for i, id := range operationIDs {
 			operationIDs[i] = "operation-id=" + id
@@ -538,82 +538,6 @@ func TestStatusRetryHandler_AttachRoutes(t *testing.T) {
 		assert.Equal(t, orchestration.Succeeded, string(op.State))
 	})
 
-	t.Run("retry failed kyma orchestration with specified operations", func(t *testing.T) {
-		// given
-		db := storage.NewMemoryStorage()
-
-		orchestrationID := "orchestration-" + fixID
-		operationIDs := []string{"id-0", "id-1", "id-2", "id-3", "id-10"}
-		err := db.Orchestrations().Insert(internal.Orchestration{OrchestrationID: orchestrationID, State: orchestration.Failed, Type: orchestration.UpgradeKymaOrchestration})
-		require.NoError(t, err)
-
-		err = fixFailedOrchestrationOperations(db, orchestrationID, orchestration.UpgradeKymaOrchestration)
-		require.NoError(t, err)
-
-		// same instance but different same type newer operation
-		err = db.Orchestrations().Insert(internal.Orchestration{OrchestrationID: "Orchestration-id-4", State: orchestration.Failed, Type: orchestration.UpgradeKymaOrchestration})
-		require.NoError(t, err)
-		sameInstOp := fixture.FixUpgradeKymaOperation("id-4", "instance-id-0")
-		sameInstOp.CreatedAt = time.Now().Add(time.Hour * 2)
-		sameInstOp.State = orchestration.Failed
-		err = db.Operations().InsertUpgradeKymaOperation(sameInstOp)
-		require.NoError(t, err)
-
-		logs := logrus.New()
-		kymaQueue := process.NewQueue(&testExecutor{}, logs)
-		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), kymaQueue, nil, 100, logs)
-
-		for i, id := range operationIDs {
-			operationIDs[i] = "operation-id=" + id
-		}
-		req, err := http.NewRequest("POST", fmt.Sprintf("/orchestrations/%s/retry", orchestrationID), strings.NewReader(strings.Join(operationIDs, "&")))
-		require.NoError(t, err)
-		req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-
-		rr := httptest.NewRecorder()
-		router := mux.NewRouter()
-		kymaHandler.AttachRoutes(router)
-
-		// when
-		router.ServeHTTP(rr, req)
-
-		// then
-		require.Equal(t, http.StatusAccepted, rr.Code)
-
-		var out orchestration.RetryResponse
-		expectedOut := orchestration.RetryResponse{
-			OrchestrationID:   orchestrationID,
-			RetryShoots:       []string{"Shoot-instance-id-2"},
-			OldOperations:     []string{"id-0"},
-			InvalidOperations: []string{"id-1", "id-3", "id-10"},
-			Msg:               "retry operations are queued for processing",
-		}
-
-		err = json.Unmarshal(rr.Body.Bytes(), &out)
-		require.NoError(t, err)
-		assert.Equal(t, expectedOut, out)
-
-		o, err := db.Orchestrations().GetByID(orchestrationID)
-		require.NoError(t, err)
-		assert.Equal(t, orchestration.Retrying, o.State)
-
-		op, err := db.Operations().GetOperationByID("id-0")
-		require.NoError(t, err)
-		assert.Equal(t, orchestration.Failed, string(op.State))
-
-		op, err = db.Operations().GetOperationByID("id-1")
-		require.NoError(t, err)
-		assert.Equal(t, orchestration.Succeeded, string(op.State))
-
-		op, err = db.Operations().GetOperationByID("id-2")
-		require.NoError(t, err)
-		assert.Equal(t, orchestration.Failed, string(op.State))
-
-		op, err = db.Operations().GetOperationByID("id-3")
-		require.NoError(t, err)
-		assert.Equal(t, orchestration.Succeeded, string(op.State))
-	})
-
 	t.Run("retry failed cluster orchestration without specified operations", func(t *testing.T) {
 		// given
 		db := storage.NewMemoryStorage()
@@ -638,7 +562,7 @@ func TestStatusRetryHandler_AttachRoutes(t *testing.T) {
 
 		logs := logrus.New()
 		clusterQueue := process.NewQueue(&testExecutor{}, logs)
-		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, clusterQueue, 100, logs)
+		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), clusterQueue, 100, logs)
 
 		req, err := http.NewRequest("POST", fmt.Sprintf("/orchestrations/%s/retry", orchestrationID), nil)
 		require.NoError(t, err)
@@ -721,76 +645,7 @@ func TestStatusRetryHandler_AttachRoutes(t *testing.T) {
 
 		logs := logrus.New()
 		clusterQueue := process.NewQueue(&testExecutor{}, logs)
-		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, clusterQueue, 100, logs)
-
-		req, err := http.NewRequest("POST", fmt.Sprintf("/orchestrations/%s/retry", orchestrationID), nil)
-		require.NoError(t, err)
-		req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-
-		rr := httptest.NewRecorder()
-		router := mux.NewRouter()
-		kymaHandler.AttachRoutes(router)
-
-		// when
-		router.ServeHTTP(rr, req)
-
-		// then
-		require.Equal(t, http.StatusAccepted, rr.Code)
-
-		var out orchestration.RetryResponse
-		expectedOut := orchestration.RetryResponse{
-			OrchestrationID:   orchestrationID,
-			RetryShoots:       []string{"Shoot-instance-id-0", "Shoot-instance-id-2"},
-			OldOperations:     nil,
-			InvalidOperations: nil,
-			Msg:               "retry operations are queued for processing",
-		}
-
-		err = json.Unmarshal(rr.Body.Bytes(), &out)
-		require.NoError(t, err)
-		assert.Equal(t, expectedOut, out)
-
-		o, err := db.Orchestrations().GetByID(orchestrationID)
-		require.NoError(t, err)
-		assert.Equal(t, orchestration.Retrying, o.State)
-
-		op, err := db.Operations().GetOperationByID("id-0")
-		require.NoError(t, err)
-		assert.Equal(t, orchestration.Failed, string(op.State))
-
-		op, err = db.Operations().GetOperationByID("id-1")
-		require.NoError(t, err)
-		assert.Equal(t, orchestration.Succeeded, string(op.State))
-
-		op, err = db.Operations().GetOperationByID("id-2")
-		require.NoError(t, err)
-		assert.Equal(t, orchestration.Failed, string(op.State))
-
-		op, err = db.Operations().GetOperationByID("id-3")
-		require.NoError(t, err)
-		assert.Equal(t, orchestration.Succeeded, string(op.State))
-	})
-
-	t.Run("retry failed kyma orchestration with deprovisioned instance", func(t *testing.T) {
-		// given
-		db := storage.NewMemoryStorage()
-
-		orchestrationID := "orchestration-" + fixID
-		err := db.Orchestrations().Insert(internal.Orchestration{OrchestrationID: orchestrationID, State: orchestration.Failed, Type: orchestration.UpgradeKymaOrchestration})
-		require.NoError(t, err)
-
-		err = fixFailedOrchestrationOperations(db, orchestrationID, orchestration.UpgradeKymaOrchestration)
-		require.NoError(t, err)
-
-		// insert a deprovisioned instance
-		deprovisioningOperation := fixture.FixDeprovisioningOperation("id-5", "instance-id-0")
-		deprovisioningOperation.State = orchestration.InProgress
-		err = db.Operations().InsertDeprovisioningOperation(deprovisioningOperation)
-		require.NoError(t, err)
-
-		logs := logrus.New()
-		kymaQueue := process.NewQueue(&testExecutor{}, logs)
-		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), kymaQueue, nil, 100, logs)
+		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), clusterQueue, 100, logs)
 
 		req, err := http.NewRequest("POST", fmt.Sprintf("/orchestrations/%s/retry", orchestrationID), nil)
 		require.NoError(t, err)
@@ -853,7 +708,7 @@ func TestStatusRetryHandler_AttachRoutes(t *testing.T) {
 
 		logs := logrus.New()
 		clusterQueue := process.NewQueue(&testExecutor{}, logs)
-		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), nil, clusterQueue, 100, logs)
+		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), clusterQueue, 100, logs)
 
 		req, err := http.NewRequest("POST", fmt.Sprintf("/orchestrations/%s/retry", orchestrationID), nil)
 		require.NoError(t, err)
@@ -901,63 +756,6 @@ func TestStatusRetryHandler_AttachRoutes(t *testing.T) {
 		op, err = db.Operations().GetOperationByID("id-3")
 		require.NoError(t, err)
 		assert.Equal(t, orchestration.Succeeded, string(op.State))
-	})
-
-	t.Run("retry failed kyma orchestration - testing scheduling", func(t *testing.T) {
-		// given
-		db := storage.NewMemoryStorage()
-
-		orchestrationID := "orchestration-" + fixID
-		operationIDs := []string{"id-2"}
-		err := db.Orchestrations().Insert(internal.Orchestration{OrchestrationID: orchestrationID, State: orchestration.Failed, Type: orchestration.UpgradeKymaOrchestration, Parameters: orchestration.Parameters{Strategy: orchestration.StrategySpec{Schedule: time.Now().Format(time.RFC3339), MaintenanceWindow: true}}})
-		require.NoError(t, err)
-
-		err = fixFailedOrchestrationOperations(db, orchestrationID, orchestration.UpgradeKymaOrchestration)
-		require.NoError(t, err)
-
-		logs := logrus.New()
-		kymaQueue := process.NewQueue(&testExecutor{}, logs)
-		kymaHandler := NewOrchestrationStatusHandler(db.Operations(), db.Orchestrations(), db.RuntimeStates(), kymaQueue, nil, 100, logs)
-
-		for i, id := range operationIDs {
-			operationIDs[i] = "operation-id=" + id
-		}
-		str := strings.Join(operationIDs, "&")
-		req, err := http.NewRequest("POST", fmt.Sprintf("/orchestrations/%s/retry", orchestrationID), strings.NewReader(str+"&immediate=true"))
-		require.NoError(t, err)
-		req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-
-		rr := httptest.NewRecorder()
-		router := mux.NewRouter()
-		kymaHandler.AttachRoutes(router)
-
-		// when
-		router.ServeHTTP(rr, req)
-
-		// then
-		require.Equal(t, http.StatusAccepted, rr.Code)
-
-		var out orchestration.RetryResponse
-		expectedOut := orchestration.RetryResponse{
-			OrchestrationID:   orchestrationID,
-			RetryShoots:       []string{"Shoot-instance-id-2"},
-			OldOperations:     nil,
-			InvalidOperations: nil,
-			Msg:               "retry operations are queued for processing",
-		}
-
-		err = json.Unmarshal(rr.Body.Bytes(), &out)
-		require.NoError(t, err)
-		assert.Equal(t, expectedOut, out)
-
-		op, err := db.Operations().GetUpgradeKymaOperationByID("id-2")
-		require.NoError(t, err)
-		assert.NotEqual(t, op.MaintenanceWindowBegin, time.Time{})
-
-		o, err := db.Orchestrations().GetByID(orchestrationID)
-		require.NoError(t, err)
-		assert.Equal(t, orchestration.Retrying, o.State)
-
 	})
 }
